flask/requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cssselect==1.2.0
Flask==2.3.2
Flask-Cors==3.0.10
frozenlist==1.3.3
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
keyboard==0.13.5
lxml==4.9.2
MarkupSafe==2.1.2
multidict==6.0.4
openai==0.27.5
pkg_resources==0.0.0
psutil==5.9.5
py-trees==2.2.3
pydot==1.4.2
pyparsing==3.0.9
readability-lxml==0.8.1
requests==2.29.0
soupsieve==2.4.1
tqdm==4.65.0
urllib3==1.26.15
Werkzeug==2.3.3
yarl==1.9.2
zipp==3.15.0
